+++ packages/sdk/src/plt/Token.ts
@@ -10,9 +10,11 @@ import {
     TokenAddAllowListOperation,
     TokenAddDenyListOperation,
     TokenAmount,
+    TokenAssignAdminRolesOperation,
     TokenBurnOperation,
     TokenId,
     TokenInfo,
+    TokenMetadataUrl,
     TokenMintOperation,
     TokenModuleReference,
     TokenModuleState,
@@ -21,9 +23,12 @@ import {
     TokenPauseOperation,
     TokenRemoveAllowListOperation,
     TokenRemoveDenyListOperation,
+    TokenRevokeAdminRolesOperation,
     TokenTransfer,
     TokenTransferOperation,
     TokenUnpauseOperation,
+    TokenUpdateAdminRolesDetails,
+    TokenUpdateMetadataOperation,
     createTokenUpdatePayload,
 } from './index.js';
 
@@ -899,6 +904,69 @@ export async function unpause(
     metadata?: TokenUpdateMetadata
 ): Promise<TransactionHash.Type> {
     const operation: TokenUnpauseOperation = { [TokenOperationType.Unpause]: {} };
+    return sendOperations(token, sender, [operation], signer, metadata);
+}
+
+/**
+ * Updates the metadata URL of a token.
+ *
+ * @param {Token} token - The token to update.
+ * @param {TokenMetadataUrl.Type} metadataUrl - The new metadata URL.
+ * @param {AccountAddress.Type} sender - The account address of the sender.
+ * @param {AccountSigner} signer - The signer responsible for signing the transaction.
+ * @param {TokenUpdateMetadata} [metadata={ expiry: TransactionExpiry.futureMinutes(5) }] - The metadata for the token update.
+ * @returns A promise that resolves to the transaction hash.
+ */
+export function updateMetadata(
+    token: Token,
+    metadataUrl: TokenMetadataUrl.Type,
+    sender: AccountAddress.Type,
+    signer: AccountSigner,
+    metadata?: TokenUpdateMetadata
+): Promise<TransactionHash.Type> {
+    const operation: TokenUpdateMetadataOperation = { [TokenOperationType.UpdateMetadata]: metadataUrl };
+    return sendOperations(token, sender, [operation], signer, metadata);
+}
+
+/**
+ * Assigns admin roles to a token.
+ *
+ * @param {Token} token - The token to update.
+ * @param {TokenUpdateAdminRolesDetails} updateAdminRoleDetails - The details of the admin role updates.
+ * @param {AccountAddress.Type} sender - The account address of the sender.
+ * @param {AccountSigner} signer - The signer responsible for signing the transaction.
+ * @param {TokenUpdateMetadata} [metadata={ expiry: TransactionExpiry.futureMinutes(5) }] - The metadata for the token update.
+ * @returns A promise that resolves to the transaction hash.
+ */
+export function assignAdminRoles(
+    token: Token,
+    updateAdminRoleDetails: TokenUpdateAdminRolesDetails,
+    sender: AccountAddress.Type,
+    signer: AccountSigner,
+    metadata?: TokenUpdateMetadata
+): Promise<TransactionHash.Type> {
+    const operation: TokenAssignAdminRolesOperation = { [TokenOperationType.AssignAdminRoles]: updateAdminRoleDetails };
+    return sendOperations(token, sender, [operation], signer, metadata);
+}
+
+/**
+ * Revokes admin roles from a token.
+ *
+ * @param {Token} token - The token to update.
+ * @param {TokenUpdateAdminRolesDetails} updateAdminRoleDetails - The details of the admin role updates.
+ * @param {AccountAddress.Type} sender - The account address of the sender.
+ * @param {AccountSigner} signer - The signer responsible for signing the transaction.
+ * @param {TokenUpdateMetadata} [metadata={ expiry: TransactionExpiry.futureMinutes(5) }] - The metadata for the token update.
+ * @returns A promise that resolves to the transaction hash.
+ */
+export function revokeAdminRoles(
+    token: Token,
+    updateAdminRoleDetails: TokenUpdateAdminRolesDetails,
+    sender: AccountAddress.Type,
+    signer: AccountSigner,
+    metadata?: TokenUpdateMetadata
+): Promise<TransactionHash.Type> {
+    const operation: TokenRevokeAdminRolesOperation = { [TokenOperationType.RevokeAdminRoles]: updateAdminRoleDetails };
     return sendOperations(token, sender, [operation], signer, metadata);
 }